refact: rename constant priorites
